+++ class/init.php
@@ -1,0 +1,57 @@
+<?php
+/*************************************************
+ * 初期設定ファイル
+ *
+ */
+
+//----------------------------------------------------
+// デバッグ表示 true / デバッグ表示オフfalse
+//----------------------------------------------------
+
+// define("_DEBUG_MODE", true); 
+
+define("_DEBUG_MODE", false); 
+
+//----------------------------------------------------
+// データベース関連
+//----------------------------------------------------
+
+// データベース接続ユーザー名
+define("_DB_USER", "root");
+
+// データベース接続パスワード
+define("_DB_PASS", "");
+
+// データベースホスト名
+define("_DB_HOST", "localhost");
+
+// データベース名
+define("_DB_NAME", "auth");
+
+// データベースの種類
+define("_DB_TYPE", "mysql");
+
+// データソースネーム
+define("_DSN", _DB_TYPE . ":host=" . _DB_HOST . ";dbname=" . _DB_NAME. ";charset=utf8");
+
+
+//----------------------------------------------------
+// ファイル設置ディレクトリ
+//----------------------------------------------------
+
+// 関連ファイルを設置するディレクトリ
+//define( "_PHP_LIBS_DIR", _ROOT_DIR . "../php_libs/");
+
+// クラスファイル
+define( "_CLASS_DIR", /*_PHP_LIBS_DIR .*/ "class/");
+
+// 環境変数 
+define( "_SCRIPT_NAME", $_SERVER['SCRIPT_NAME']);
+
+
+//----------------------------------------------------
+// クラスファイルの読み込み
+//----------------------------------------------------
+// 読み込みの順番を変えると動作しません。
+//require_once(_CLASS_DIR . "BaseModel.php");
+?>
